+++ CHANGELOG.md
@@ -189,4 +189,4 @@ const localizer = momentLocalizer(moment);
 ## [1.0.0] 2017-09-20
 ### Original Release
 - Added React-Bootstrap as base framework
-- Added design from Light Bootstrap Dashboard by Creative Tim
+- Added design from AMI | Dashboard by Creative Tim